remove duplicated kubeconfig loading methods
